+++ Litet Spelprojekt/Game/Source/GUI/UIReplay.cpp
@@ -28,7 +28,7 @@ void UIReplay::RenderProgress(GUIContext* context, float x, float y) noexcept
 	stream << ReplayHandler::GetElapsedTime() << " s";
 	static float width = 6.0F;
 	Window& pWindow = Game::GetGame()->GetWindow();
-	context->RenderText(stream.str(), GetWidth() * GetPercentage() - 20, y + GetHeight() * 2.0F, pWindow.GetWidth(), pWindow.GetHeight(), 0.5F, GUIContext::COLOR_WHITE, CENTER_VERTICAL);
+	context->RenderText(stream.str(), GetWidth() * GetPercentage() - 20, y + GetHeight() * 2.0F, (float)pWindow.GetWidth(), (float)pWindow.GetHeight(), 0.5F, GUIContext::COLOR_WHITE, CENTER_VERTICAL);
 	context->RenderTexture(GetDefaultTexture(), GetWidth() * GetPercentage() - width / 2.0F, y - GetHeight() / 2.0F, width, GetHeight() * 2, GUIContext::COLOR_WHITE);
 }
